Revert "O Morozova - Settings ViewModel"
