--- src/Data/DAL/Setup/IMillPipeSizeTypeRepository.cs
@@ -1,13 +1,0 @@
-﻿using Domain.Entity.Setup;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace Data.DAL.Setup
-{
-    public interface IMillPipeSizeTypeRepository : IRepository<Guid, PipeMillSizeType>
-    {
-    }
-}
