+++ src/Data/Data.csproj
@@ -50,22 +50,16 @@
   <ItemGroup>
     <Compile Include="DAL\Hibernate\AbstractHibernateRepository.cs" />
     <Compile Include="DAL\Hibernate\HibernateUtil.cs" />
-    <Compile Include="DAL\Hibernate\MillPipeSizeTypeRepository.cs" />
-    <Compile Include="DAL\Hibernate\PipeTestRepository.cs" />
     <Compile Include="DAL\Hibernate\PipeRepository.cs" />
     <Compile Include="DAL\Hibernate\RailcarRepository.cs" />
     <Compile Include="DAL\IRepository.cs" />
     <Compile Include="DAL\Mapping\ItemMap.cs" />
     <Compile Include="DAL\Mapping\PipelinePieceMap.cs" />
     <Compile Include="DAL\Mapping\PipeMap.cs" />
-    <Compile Include="DAL\Mapping\PipeMillSizeTypeMap.cs" />
-    <Compile Include="DAL\Mapping\PipeTestMap.cs" />
     <Compile Include="DAL\Mapping\PurchaseOrderMap.cs" />
     <Compile Include="DAL\Mapping\RailcarMap.cs" />
     <Compile Include="DAL\Mill\IPipeRepository.cs" />
     <Compile Include="DAL\Mill\IRailcarRepository.cs" />
-    <Compile Include="DAL\Setup\IMillPipeSizeTypeRepository.cs" />
-    <Compile Include="DAL\Setup\IPipeTestRepository.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
   </ItemGroup>
   <ItemGroup>
