+++ src/Domain/Entity/Mill/PipeTestControlType.cs
@@ -1,0 +1,12 @@
+﻿namespace Domain.Entity.Mill
+{
+    public enum PipeTestControlType
+    {
+        Witness = 1,
+        Review = 2,
+        Monitor = 3,
+        Hold = 4,
+
+        Undef = 0
+    }
+}
