+++ src/Domain/Entity/Mill/PipeTestResult.cs
@@ -1,5 +1,4 @@
-﻿using Domain.Entity.Setup;
-using System;
+﻿using System;
 using System.Collections.Generic;
 
 namespace Domain.Entity.Mill
