+++ src/Domain/Entity/Mill/PipeTestResultType.cs
@@ -1,0 +1,11 @@
+﻿namespace Domain.Entity.Mill
+{
+    public enum PipeTestResultType
+    {
+        Boolean = 1,
+        String = 2,
+        Diapason = 3,
+
+        Undef = 0
+    }
+}
