--- src/Domain/Entity/Setup/PipeMillSizeType.cs
@@ -1,15 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace Domain.Entity.Setup
-{
-    public class PipeMillSizeType: Item
-    {
-        public virtual string Name { get; set; }
-        public virtual IList<PipeTest> PipeTests { get; set; }
-
-    }
-}
