--- src/Domain/Entity/Setup/PipeTestControlType.cs
@@ -1,18 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace Domain.Entity.Setup
-{
-    public enum PipeTestControlType
-    {
-        Witness = 1,
-        Review = 2,
-        Monitor = 3,
-        Hold = 4,
-
-        Undef = 0
-    }
-}
