--- src/Domain/Entity/Setup/PipeTestResultType.cs
@@ -1,17 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace Domain.Entity.Setup
-{
-    public enum PipeTestResultType
-    {
-        Boolean = 1,
-        String = 2,
-        Diapason = 3,
-
-        Undef = 0
-    }
-}
